+++ Runtime/Utilities/InventorySystem/Inventories/ClassicInventory.cs
@@ -4,18 +4,6 @@ namespace SkelTech.RPEST.Utilities.InventorySystem {
         protected override ClassicItem<T> CreateItem(T itemData, int amount) {
             return new ClassicItem<T>(itemData, amount);
         }
-
-        protected override bool NewItemCondition(T itemData, int amount) {
-            return amount <= itemData.MaximumCount;
-        }
-
-        protected override bool IncrementItemCondition(ClassicItem<T> item, int amount) {
-            return (item.Count + amount) <= item.ItemData.MaximumCount;
-        }
-
-        protected override void ClearItem(ClassicItem<T> item) {
-            this.items.Remove(item.ItemData.Id);
-        }
         #endregion
     }
 }
