+++ Runtime/Utilities/InventorySystem/ItemData/ClassicItemData.cs
@@ -1,5 +1,18 @@
 using UnityEngine;
 
 namespace SkelTech.RPEST.Utilities.InventorySystem {
-    public class ClassicItemData : ItemData {}
+    public class ClassicItemData : ScriptableObject {
+        #region Properties
+        public int Id { get { return this.id; }}
+        public string ItemName { get { return this.itemName; }}
+        public int MaximumCount { get { return this.maximumCount; }}
+        #endregion
+
+        #region Fields
+        [SerializeField] private int id;
+        [SerializeField] private string itemName;
+        [SerializeField, Min(1)] private int maximumCount = 999999999;
+        
+        #endregion
+    }
 }
