+++ Runtime/Utilities/InventorySystem/ItemData/ClassicItemData.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2a1c095b20769f54586e17d90c258f54
+guid: 7d5df4acac443524f9090eef9cdc45d5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
