--- Runtime/Utilities/InventorySystem/ItemData/ItemData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7d5df4acac443524f9090eef9cdc45d5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
