+++ Runtime/Utilities/InventorySystem/ItemData/WeightedItemData.cs
@@ -1,13 +1,13 @@
 using UnityEngine;
 
 namespace SkelTech.RPEST.Utilities.InventorySystem {
-    public class WeightedItemData : ItemData {
+    public class WeightedItemData : ClassicItemData {
         #region Properties
         public float Weight { get { return this.weight; }}
         #endregion
 
         #region Fields
-        [SerializeField] private float weight;
+        [SerializeField, Min(0)] private float weight;
         #endregion
     }
 }
