+++ Runtime/Utilities/InventorySystem/Items/ClassicItem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7162e8d38a57502469d9aadf05a81812
+guid: 169e079795bce2d4fa23f7ef4f7f183b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
