--- Runtime/Utilities/InventorySystem/Items/Item.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 169e079795bce2d4fa23f7ef4f7f183b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
