+++ Runtime/Utilities/InventorySystem/Items/WeightedItem.cs
@@ -1,5 +1,5 @@
 namespace SkelTech.RPEST.Utilities.InventorySystem {
-    public class WeightedItem<T> : Item<T> where T : WeightedItemData {
+    public class WeightedItem<T> : ClassicItem<T> where T : WeightedItemData {
         #region Constructors
         public WeightedItem(T itemData) : base(itemData) {}
         public WeightedItem(T itemData, int count) : base(itemData, count) {}
